Attendance would be marked online, so this would help in minimizing paper cost and images would be real time captured so there would be regular classes for the students. It can work on desktop and also on Mobiles.
According to the survey we performed in different Government Schools in Rajasthan, we came to know that the teachers are not taking classes regularly but are paid full salary.
Here we design a system by which Face Recognisation would be done by taking current picture of them in class and matching them with their pictures in database.
this will in-turn help in keeping long book records and all the data would be online.